Rename file to are_lists_equal
